+++ Voxel2PixelTest/ImageMaker.cs
@@ -8,7 +8,7 @@ namespace Voxel2PixelTest
 {
 	public static class ImageMaker
 	{
-		public static byte[][] AddFrameNumbers(this byte[][] frames, int width = 0, uint color = 0xFFFFFFFF)
+		public static byte[][] AddFrameNumbers(this byte[][] frames, ushort width = 0, uint color = 0xFFFFFFFF)
 		{
 			for (int frame = 0; frame < frames.Length; frame++)
 				frames[frame].Draw3x4(
@@ -28,29 +28,29 @@ namespace Voxel2PixelTest
 				width: width,
 				height: bytes.Length / width >> 2);
 		}
-		public static SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> Png(int scaleX, int scaleY, int width = 0, params byte[] bytes)
+		public static SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> Png(ushort scaleX, ushort scaleY, ushort width = 0, params byte[] bytes)
 		{
 			if (scaleX == 1 && scaleY == 1)
 				return Png(width, bytes);
 			if (width < 1)
-				width = (int)Math.Sqrt(bytes.Length >> 2);
+				width = (ushort)Math.Sqrt(bytes.Length >> 2);
 			return Image.LoadPixelData<SixLabors.ImageSharp.PixelFormats.Rgba32>(
 				data: bytes.Upscale(scaleX, scaleY, width),
 				width: width * scaleX,
 				height: (bytes.Length / width >> 2) * scaleY);
 		}
-		public static SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> AnimatedGif(int scaleX, int scaleY, int width = 0, int frameDelay = 25, ushort repeatCount = 0, params byte[][] frames) => AnimatedGif(
-			width: width * scaleX,
+		public static SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> AnimatedGif(ushort scaleX, ushort scaleY, ushort width = 0, int frameDelay = 25, ushort repeatCount = 0, params byte[][] frames) => AnimatedGif(
+			width: (ushort)(width * scaleX),
 			frameDelay: frameDelay,
 			repeatCount: repeatCount,
 			frames: scaleX == 1 && scaleY == 1 ? frames
 				: frames
 					.Select(f => f.Upscale(scaleX, scaleY, width))
 					.ToArray());
-		public static SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> AnimatedGif(int width = 0, int frameDelay = 25, ushort repeatCount = 0, params byte[][] frames)
+		public static SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> AnimatedGif(ushort width = 0, int frameDelay = 25, ushort repeatCount = 0, params byte[][] frames)
 		{
 			if (width < 1)
-				width = (int)Math.Sqrt(frames[0].Length >> 2);
+				width = (ushort)Math.Sqrt(frames[0].Length >> 2);
 			int height = frames[0].Length / width >> 2;
 			SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32> gif = new SixLabors.ImageSharp.Image<SixLabors.ImageSharp.PixelFormats.Rgba32>(width, height);
 			SixLabors.ImageSharp.Formats.Gif.GifMetadata gifMetaData = gif.Metadata.GetGifMetadata();
