+++ Voxel2PixelTest/Model/TextureModelTest.cs
@@ -45,7 +45,7 @@ namespace Voxel2PixelTest.Model
 				0,255,0,255,
 				0,0,255,255,
 				128,128,128,255}
-			.Upscale(width / 2, height / 2, 2)
+			.Upscale((ushort)(width / 2), (ushort)(height / 2), 2)
 			.DrawRectangle(0, 0, 0, 255, width / 4, height / 4, width / 2, height / 2, width);
 	}
 }
